Déplacement des trucs de dépots dans leur propre package
