+++ Source/CodeNameHippie/app/src/main/java/com/pam/codenamehippie/ui/ListeMarchandisesDisponiblesActivity.java
@@ -7,9 +7,9 @@ import android.widget.ExpandableListView;
 import com.pam.codenamehippie.HippieApplication;
 import com.pam.codenamehippie.R;
 import com.pam.codenamehippie.modele.AlimentaireModele;
-import com.pam.codenamehippie.modele.AlimentaireModeleDepot;
-import com.pam.codenamehippie.modele.FiltreDeListe;
-import com.pam.codenamehippie.modele.ObservateurDeDepot;
+import com.pam.codenamehippie.modele.depot.AlimentaireModeleDepot;
+import com.pam.codenamehippie.modele.depot.FiltreDeListe;
+import com.pam.codenamehippie.modele.depot.ObservateurDeDepot;
 import com.pam.codenamehippie.ui.adapter.ListeMarchandisesDisponiblesAdapter;
 
 import java.io.IOException;
