+++ Source/CodeNameHippie/app/src/main/java/com/pam/codenamehippie/ui/ListeMesDonsActivity.java
@@ -10,9 +10,9 @@ import android.widget.Toast;
 import com.pam.codenamehippie.HippieApplication;
 import com.pam.codenamehippie.R;
 import com.pam.codenamehippie.modele.AlimentaireModele;
-import com.pam.codenamehippie.modele.AlimentaireModeleDepot;
-import com.pam.codenamehippie.modele.FiltreDeListe;
-import com.pam.codenamehippie.modele.ObservateurDeDepot;
+import com.pam.codenamehippie.modele.depot.AlimentaireModeleDepot;
+import com.pam.codenamehippie.modele.depot.FiltreDeListe;
+import com.pam.codenamehippie.modele.depot.ObservateurDeDepot;
 import com.pam.codenamehippie.ui.adapter.HippieListAdapter;
 
 import java.io.IOException;
