+++ Source/CodeNameHippie/app/src/main/java/com/pam/codenamehippie/ui/adapter/HippieListAdapter.java
@@ -16,7 +16,7 @@ import android.widget.Toast;
 
 import com.pam.codenamehippie.R;
 import com.pam.codenamehippie.modele.AlimentaireModele;
-import com.pam.codenamehippie.modele.AlimentaireModeleDepot;
+import com.pam.codenamehippie.modele.depot.AlimentaireModeleDepot;
 import com.pam.codenamehippie.ui.AjoutMarchandiseActivity;
 
 import java.text.DateFormat;
